feat(model): suport async model trigger (#543)
Because

- we are going to support asynchronous model trigger

This commit

- support async model trigger with temporal workflow
