+++ pkg/ray/postprocess.go
@@ -3,15 +3,22 @@ package ray
 import (
 	"bytes"
 	"encoding/base64"
+	"encoding/json"
 	"fmt"
 	"image"
 	"image/color"
 	"image/jpeg"
 	"log"
+	"strconv"
+	"strings"
+
+	"google.golang.org/protobuf/encoding/protojson"
+	"google.golang.org/protobuf/types/known/structpb"
 
 	"github.com/instill-ai/model-backend/pkg/ray/rayserver"
-	// "github.com/instill-ai/model-backend/pkg/ray"
+	"github.com/instill-ai/model-backend/pkg/utils"
 	commonPB "github.com/instill-ai/protogen-go/common/task/v1alpha"
+	modelPB "github.com/instill-ai/protogen-go/model/model/v1alpha"
 )
 
 func postProcessDetection(modelInferResponse *rayserver.RayServiceCallResponse, outputNameBboxes string, outputNameLabels string) (any, error) {
@@ -525,5 +532,534 @@ func postProcessTextGeneration(modelInferResponse *rayserver.RayServiceCallRespo
 		return TextGenerationOutput{
 			Text: outputTexts,
 		}, nil
+	}
+}
+
+func PostProcess(inferResponse *rayserver.RayServiceCallResponse, modelMetadata *rayserver.ModelMetadataResponse, task commonPB.Task) (outputs []*modelPB.TaskOutput, err error) {
+
+	var postprocessResponse any
+
+	switch task {
+	case commonPB.Task_TASK_CLASSIFICATION:
+		postprocessResponse, err = postProcessClassification(inferResponse, modelMetadata.Outputs[0].Name)
+		if err != nil {
+			return nil, fmt.Errorf("unable to post-process classification output: %w", err)
+		}
+		clsResponses := postprocessResponse.([]string)
+		var clsOutputs []*modelPB.TaskOutput
+		for _, clsRes := range clsResponses {
+			clsResSplit := strings.Split(clsRes, ":")
+			if len(clsResSplit) == 2 {
+				score, err := strconv.ParseFloat(clsResSplit[0], 32)
+				if err != nil {
+					return nil, fmt.Errorf("unable to decode inference output")
+				}
+				clsOutput := modelPB.TaskOutput{
+					Output: &modelPB.TaskOutput_Classification{
+						Classification: &modelPB.ClassificationOutput{
+							Category: clsResSplit[1],
+							Score:    float32(score),
+						},
+					},
+				}
+				clsOutputs = append(clsOutputs, &clsOutput)
+			} else if len(clsResSplit) == 3 {
+				score, err := strconv.ParseFloat(clsResSplit[0], 32)
+				if err != nil {
+					return nil, fmt.Errorf("unable to decode inference output")
+				}
+				clsOutput := modelPB.TaskOutput{
+					Output: &modelPB.TaskOutput_Classification{
+						Classification: &modelPB.ClassificationOutput{
+							Category: clsResSplit[2],
+							Score:    float32(score),
+						},
+					},
+				}
+				clsOutputs = append(clsOutputs, &clsOutput)
+			} else {
+				return nil, fmt.Errorf("unable to decode inference output")
+			}
+		}
+		if len(clsOutputs) == 0 {
+			clsOutputs = append(clsOutputs, &modelPB.TaskOutput{
+				Output: &modelPB.TaskOutput_Classification{
+					Classification: &modelPB.ClassificationOutput{},
+				},
+			})
+		}
+		return clsOutputs, nil
+	case commonPB.Task_TASK_DETECTION:
+		if len(modelMetadata.Outputs) < 2 {
+			return nil, fmt.Errorf("wrong output format of detection task")
+		}
+		postprocessResponse, err = postProcessDetection(inferResponse, modelMetadata.Outputs[0].Name, modelMetadata.Outputs[1].Name)
+		if err != nil {
+			return nil, fmt.Errorf("unable to post-process detection output: %w", err)
+		}
+		detResponses := postprocessResponse.(DetectionOutput)
+		batchedOutputDataBboxes := detResponses.Boxes
+		batchedOutputDataLabels := detResponses.Labels
+		var detOutputs []*modelPB.TaskOutput
+		for i := range batchedOutputDataBboxes {
+			var detOutput = modelPB.TaskOutput{
+				Output: &modelPB.TaskOutput_Detection{
+					Detection: &modelPB.DetectionOutput{
+						Objects: []*modelPB.DetectionObject{},
+					},
+				},
+			}
+			for j := range batchedOutputDataBboxes[i] {
+				box := batchedOutputDataBboxes[i][j]
+				label := batchedOutputDataLabels[i][j]
+				// Non-meaningful bboxes were added with coords [-1, -1, -1, -1, -1] and label "0" for Ray to be able to batch Tensors
+				if label != "0" {
+					bbObj := &modelPB.DetectionObject{
+						Category: label,
+						Score:    box[4],
+						// Convert x1y1x2y2 to xywh where xy is top-left corner
+						BoundingBox: &modelPB.BoundingBox{
+							Left:   box[0],
+							Top:    box[1],
+							Width:  box[2] - box[0],
+							Height: box[3] - box[1],
+						},
+					}
+					detOutput.GetDetection().Objects = append(detOutput.GetDetection().Objects, bbObj)
+				}
+			}
+			detOutputs = append(detOutputs, &detOutput)
+		}
+		if len(detOutputs) == 0 {
+			detOutputs = append(detOutputs, &modelPB.TaskOutput{
+				Output: &modelPB.TaskOutput_Detection{
+					Detection: &modelPB.DetectionOutput{
+						Objects: []*modelPB.DetectionObject{},
+					},
+				},
+			})
+		}
+		return detOutputs, nil
+	case commonPB.Task_TASK_KEYPOINT:
+		if len(modelMetadata.Outputs) < 3 {
+			return nil, fmt.Errorf("wrong output format of keypoint detection task")
+		}
+		postprocessResponse, err = postProcessKeypoint(inferResponse, modelMetadata.Outputs[0].Name, modelMetadata.Outputs[1].Name, modelMetadata.Outputs[2].Name)
+		if err != nil {
+			return nil, fmt.Errorf("unable to post-process keypoint output: %w", err)
+		}
+		keypointResponse := postprocessResponse.(KeypointOutput)
+		var keypointOutputs []*modelPB.TaskOutput
+		for i := range keypointResponse.Keypoints { // batch size
+			var keypointObjects []*modelPB.KeypointObject
+			for j := range keypointResponse.Keypoints[i] { // n keypoints in one image
+				if keypointResponse.Scores[i][j] == -1 { // dummy object for batching to make sure every images have same output shape
+					continue
+				}
+				var keypoints []*modelPB.Keypoint
+				points := keypointResponse.Keypoints[i][j]
+				for k := range points { // 17 point for each keypoint
+					if points[k][0] == -1 && points[k][1] == -1 && points[k][2] == -1 { // dummy output for batching to make sure every images have same output shape
+						continue
+					}
+					keypoints = append(keypoints, &modelPB.Keypoint{
+						X: points[k][0],
+						Y: points[k][1],
+						V: points[k][2],
+					})
+				}
+				keypointObjects = append(keypointObjects, &modelPB.KeypointObject{
+					Keypoints: keypoints,
+					BoundingBox: &modelPB.BoundingBox{
+						Left:   keypointResponse.Boxes[i][j][0],
+						Top:    keypointResponse.Boxes[i][j][1],
+						Width:  keypointResponse.Boxes[i][j][2],
+						Height: keypointResponse.Boxes[i][j][3],
+					},
+					Score: keypointResponse.Scores[i][j],
+				})
+			}
+			keypointOutputs = append(keypointOutputs, &modelPB.TaskOutput{
+				Output: &modelPB.TaskOutput_Keypoint{
+					Keypoint: &modelPB.KeypointOutput{
+						Objects: keypointObjects,
+					},
+				},
+			})
+		}
+		if len(keypointOutputs) == 0 {
+			keypointOutputs = append(keypointOutputs, &modelPB.TaskOutput{
+				Output: &modelPB.TaskOutput_Keypoint{
+					Keypoint: &modelPB.KeypointOutput{
+						Objects: []*modelPB.KeypointObject{},
+					},
+				},
+			})
+		}
+		return keypointOutputs, nil
+	case commonPB.Task_TASK_OCR:
+		if len(modelMetadata.Outputs) < 2 {
+			return nil, fmt.Errorf("wrong output format of OCR task")
+		}
+		switch len(modelMetadata.Outputs) {
+		case 2:
+			postprocessResponse, err = postProcessOcrWithoutScore(inferResponse, modelMetadata.Outputs[0].Name, modelMetadata.Outputs[1].Name)
+			if err != nil {
+				return nil, fmt.Errorf("unable to post-process detection output: %w", err)
+			}
+		case 3:
+			postprocessResponse, err = postProcessOcrWithScore(inferResponse, modelMetadata.Outputs[0].Name, modelMetadata.Outputs[1].Name, modelMetadata.Outputs[2].Name)
+			if err != nil {
+				return nil, fmt.Errorf("unable to post-process detection output: %w", err)
+			}
+		}
+		ocrResponses := postprocessResponse.(OcrOutput)
+		batchedOutputDataBboxes := ocrResponses.Boxes
+		batchedOutputDataTexts := ocrResponses.Texts
+		batchedOutputDataScores := ocrResponses.Scores
+		var ocrOutputs []*modelPB.TaskOutput
+		for i := range batchedOutputDataBboxes {
+			var ocrOutput = modelPB.TaskOutput{
+				Output: &modelPB.TaskOutput_Ocr{
+					Ocr: &modelPB.OcrOutput{
+						Objects: []*modelPB.OcrObject{},
+					},
+				},
+			}
+			for j := range batchedOutputDataBboxes[i] {
+				box := batchedOutputDataBboxes[i][j]
+				text := batchedOutputDataTexts[i][j]
+				score := batchedOutputDataScores[i][j]
+				// Non-meaningful bboxes were added with coords [-1, -1, -1, -1, -1] and text "" for Ray to be able to batch Tensors
+				if text != "" && box[0] != -1 {
+					ocrOutput.GetOcr().Objects = append(ocrOutput.GetOcr().Objects, &modelPB.OcrObject{
+						BoundingBox: &modelPB.BoundingBox{
+							Left:   box[0],
+							Top:    box[1],
+							Width:  box[2],
+							Height: box[3],
+						},
+						Score: score,
+						Text:  text,
+					})
+				}
+			}
+			ocrOutputs = append(ocrOutputs, &ocrOutput)
+		}
+		if len(ocrOutputs) == 0 {
+			ocrOutputs = append(ocrOutputs, &modelPB.TaskOutput{
+				Output: &modelPB.TaskOutput_Ocr{
+					Ocr: &modelPB.OcrOutput{
+						Objects: []*modelPB.OcrObject{},
+					},
+				},
+			})
+		}
+		return ocrOutputs, nil
+	case commonPB.Task_TASK_INSTANCE_SEGMENTATION:
+		if len(modelMetadata.Outputs) < 4 {
+			return nil, fmt.Errorf("wrong output format of instance segmentation task")
+		}
+		postprocessResponse, err = postProcessInstanceSegmentation(inferResponse, modelMetadata.Outputs[0].Name, modelMetadata.Outputs[1].Name, modelMetadata.Outputs[2].Name, modelMetadata.Outputs[3].Name)
+		if err != nil {
+			return nil, fmt.Errorf("unable to post-process instance segmentation output: %w", err)
+		}
+		instanceSegmentationResponses := postprocessResponse.(InstanceSegmentationOutput)
+		batchedOutputDataRles := instanceSegmentationResponses.Rles
+		batchedOutputDataBboxes := instanceSegmentationResponses.Boxes
+		batchedOutputDataLabels := instanceSegmentationResponses.Labels
+		batchedOutputDataScores := instanceSegmentationResponses.Scores
+		var instanceSegmentationOutputs []*modelPB.TaskOutput
+		for i := range batchedOutputDataBboxes {
+			var instanceSegmentationOutput = modelPB.TaskOutput{
+				Output: &modelPB.TaskOutput_InstanceSegmentation{
+					InstanceSegmentation: &modelPB.InstanceSegmentationOutput{
+						Objects: []*modelPB.InstanceSegmentationObject{},
+					},
+				},
+			}
+			for j := range batchedOutputDataBboxes[i] {
+				rle := batchedOutputDataRles[i][j]
+				box := batchedOutputDataBboxes[i][j]
+				label := batchedOutputDataLabels[i][j]
+				score := batchedOutputDataScores[i][j]
+				// Non-meaningful bboxes were added with coords [-1, -1, -1, -1, -1] and text "" for Ray to be able to batch Tensors
+				if label != "" && rle != "" {
+					instanceSegmentationOutput.GetInstanceSegmentation().Objects = append(instanceSegmentationOutput.GetInstanceSegmentation().Objects, &modelPB.InstanceSegmentationObject{
+						Rle: rle,
+						BoundingBox: &modelPB.BoundingBox{
+							Left:   box[0],
+							Top:    box[1],
+							Width:  box[2],
+							Height: box[3],
+						},
+						Score:    score,
+						Category: label,
+					})
+				}
+			}
+			instanceSegmentationOutputs = append(instanceSegmentationOutputs, &instanceSegmentationOutput)
+		}
+		if len(instanceSegmentationOutputs) == 0 {
+			instanceSegmentationOutputs = append(instanceSegmentationOutputs, &modelPB.TaskOutput{
+				Output: &modelPB.TaskOutput_InstanceSegmentation{
+					InstanceSegmentation: &modelPB.InstanceSegmentationOutput{
+						Objects: []*modelPB.InstanceSegmentationObject{},
+					},
+				},
+			})
+		}
+		return instanceSegmentationOutputs, nil
+	case commonPB.Task_TASK_SEMANTIC_SEGMENTATION:
+		if len(modelMetadata.Outputs) < 2 {
+			return nil, fmt.Errorf("wrong output format of semantic segmentation task")
+		}
+		postprocessResponse, err = postProcessSemanticSegmentation(inferResponse, modelMetadata.Outputs[0].Name, modelMetadata.Outputs[1].Name)
+		if err != nil {
+			return nil, fmt.Errorf("unable to post-process semantic segmentation output: %w", err)
+		}
+		semanticSegmentationResponses := postprocessResponse.(SemanticSegmentationOutput)
+		batchedOutputDataRles := semanticSegmentationResponses.Rles
+		batchedOutputDataCategories := semanticSegmentationResponses.Categories
+		var semanticSegmentationOutputs []*modelPB.TaskOutput
+		for i := range batchedOutputDataCategories { // loop over images
+			var semanticSegmentationOutput = modelPB.TaskOutput{
+				Output: &modelPB.TaskOutput_SemanticSegmentation{
+					SemanticSegmentation: &modelPB.SemanticSegmentationOutput{
+						Stuffs: []*modelPB.SemanticSegmentationStuff{},
+					},
+				},
+			}
+			for j := range batchedOutputDataCategories[i] { // single image
+				rle := batchedOutputDataRles[i][j]
+				category := batchedOutputDataCategories[i][j]
+				// Non-meaningful bboxes were added with coords [-1, -1, -1, -1, -1] and text "" for Ray to be able to batch Tensors
+				if category != "" && rle != "" {
+					semanticSegmentationOutput.GetSemanticSegmentation().Stuffs = append(semanticSegmentationOutput.GetSemanticSegmentation().Stuffs, &modelPB.SemanticSegmentationStuff{
+						Rle:      rle,
+						Category: category,
+					})
+				}
+			}
+			semanticSegmentationOutputs = append(semanticSegmentationOutputs, &semanticSegmentationOutput)
+		}
+		if len(semanticSegmentationOutputs) == 0 {
+			semanticSegmentationOutputs = append(semanticSegmentationOutputs, &modelPB.TaskOutput{
+				Output: &modelPB.TaskOutput_SemanticSegmentation{
+					SemanticSegmentation: &modelPB.SemanticSegmentationOutput{
+						Stuffs: []*modelPB.SemanticSegmentationStuff{},
+					},
+				},
+			})
+		}
+		return semanticSegmentationOutputs, nil
+	case commonPB.Task_TASK_TEXT_TO_IMAGE:
+		postprocessResponse, err = postProcessTextToImage(inferResponse, modelMetadata.Outputs[0].Name, task)
+		if err != nil {
+			return nil, fmt.Errorf("unable to post-process text to image output: %w", err)
+		}
+		textToImageResponses := postprocessResponse.(TextToImageOutput)
+		batchedOutputDataImages := textToImageResponses.Images
+		var textToImageOutputs []*modelPB.TaskOutput
+		for i := range batchedOutputDataImages { // loop over images
+			var textToImageOutput = modelPB.TaskOutput{
+				Output: &modelPB.TaskOutput_TextToImage{
+					TextToImage: &modelPB.TextToImageOutput{
+						Images: batchedOutputDataImages[i],
+					},
+				},
+			}
+
+			textToImageOutputs = append(textToImageOutputs, &textToImageOutput)
+		}
+		if len(textToImageOutputs) == 0 {
+			textToImageOutputs = append(textToImageOutputs, &modelPB.TaskOutput{
+				Output: &modelPB.TaskOutput_TextToImage{
+					TextToImage: &modelPB.TextToImageOutput{
+						Images: []string{},
+					},
+				},
+			})
+		}
+		return textToImageOutputs, nil
+	case commonPB.Task_TASK_IMAGE_TO_IMAGE:
+		postprocessResponse, err = postProcessTextToImage(inferResponse, modelMetadata.Outputs[0].Name, task)
+		if err != nil {
+			return nil, fmt.Errorf("unable to post-process image to image output: %w", err)
+		}
+		imageToImageResponses := postprocessResponse.(ImageToImageOutput)
+		batchedOutputDataImages := imageToImageResponses.Images
+		var imageToImageOutputs []*modelPB.TaskOutput
+		for i := range batchedOutputDataImages { // loop over images
+			var imageToImageOutput = modelPB.TaskOutput{
+				Output: &modelPB.TaskOutput_ImageToImage{
+					ImageToImage: &modelPB.ImageToImageOutput{
+						Images: batchedOutputDataImages[i],
+					},
+				},
+			}
+
+			imageToImageOutputs = append(imageToImageOutputs, &imageToImageOutput)
+		}
+		if len(imageToImageOutputs) == 0 {
+			imageToImageOutputs = append(imageToImageOutputs, &modelPB.TaskOutput{
+				Output: &modelPB.TaskOutput_ImageToImage{
+					ImageToImage: &modelPB.ImageToImageOutput{
+						Images: []string{},
+					},
+				},
+			})
+		}
+		return imageToImageOutputs, nil
+	case commonPB.Task_TASK_TEXT_GENERATION:
+		postprocessResponse, err = postProcessTextGeneration(inferResponse, modelMetadata.Outputs[0].Name, task)
+		if err != nil {
+			return nil, fmt.Errorf("unable to post-process text generation output: %w", err)
+		}
+		textGenerationResponses := postprocessResponse.(TextGenerationOutput)
+		batchedOutputDataTexts := textGenerationResponses.Text
+		var textGenerationOutputs []*modelPB.TaskOutput
+		for i := range batchedOutputDataTexts {
+			var textGenerationOutput = modelPB.TaskOutput{
+				Output: &modelPB.TaskOutput_TextGeneration{
+					TextGeneration: &modelPB.TextGenerationOutput{
+						Text: batchedOutputDataTexts[i],
+					},
+				},
+			}
+
+			textGenerationOutputs = append(textGenerationOutputs, &textGenerationOutput)
+		}
+		if len(textGenerationOutputs) == 0 {
+			textGenerationOutputs = append(textGenerationOutputs, &modelPB.TaskOutput{
+				Output: &modelPB.TaskOutput_TextGeneration{
+					TextGeneration: &modelPB.TextGenerationOutput{
+						Text: "",
+					},
+				},
+			})
+		}
+		return textGenerationOutputs, nil
+	case commonPB.Task_TASK_VISUAL_QUESTION_ANSWERING:
+		postprocessResponse, err = postProcessTextGeneration(inferResponse, modelMetadata.Outputs[0].Name, task)
+		if err != nil {
+			return nil, fmt.Errorf("unable to post-process visual question answering output: %w", err)
+		}
+		visualQuestionAnsweringResponses := postprocessResponse.(VisualQuestionAnsweringOutput)
+		batchedOutputDataTexts := visualQuestionAnsweringResponses.Text
+		var visualQuestionAnsweringOutputs []*modelPB.TaskOutput
+		for i := range batchedOutputDataTexts {
+			var visualQuestionAnsweringOutput = modelPB.TaskOutput{
+				Output: &modelPB.TaskOutput_VisualQuestionAnswering{
+					VisualQuestionAnswering: &modelPB.VisualQuestionAnsweringOutput{
+						Text: batchedOutputDataTexts[i],
+					},
+				},
+			}
+
+			visualQuestionAnsweringOutputs = append(visualQuestionAnsweringOutputs, &visualQuestionAnsweringOutput)
+		}
+		if len(visualQuestionAnsweringOutputs) == 0 {
+			visualQuestionAnsweringOutputs = append(visualQuestionAnsweringOutputs, &modelPB.TaskOutput{
+				Output: &modelPB.TaskOutput_VisualQuestionAnswering{
+					VisualQuestionAnswering: &modelPB.VisualQuestionAnsweringOutput{
+						Text: "",
+					},
+				},
+			})
+		}
+		return visualQuestionAnsweringOutputs, nil
+	case commonPB.Task_TASK_TEXT_GENERATION_CHAT:
+		postprocessResponse, err = postProcessTextGeneration(inferResponse, modelMetadata.Outputs[0].Name, task)
+		if err != nil {
+			return nil, fmt.Errorf("unable to post-process text to text output: %w", err)
+		}
+		textGenerationChatResponses := postprocessResponse.(TextGenerationChatOutput)
+		batchedOutputDataTexts := textGenerationChatResponses.Text
+		var textGenerationChatOutputs []*modelPB.TaskOutput
+		for i := range batchedOutputDataTexts {
+			var textGenerationChatOutput = modelPB.TaskOutput{
+				Output: &modelPB.TaskOutput_TextGenerationChat{
+					TextGenerationChat: &modelPB.TextGenerationChatOutput{
+						Text: batchedOutputDataTexts[i],
+					},
+				},
+			}
+
+			textGenerationChatOutputs = append(textGenerationChatOutputs, &textGenerationChatOutput)
+		}
+		if len(textGenerationChatOutputs) == 0 {
+			textGenerationChatOutputs = append(textGenerationChatOutputs, &modelPB.TaskOutput{
+				Output: &modelPB.TaskOutput_TextGenerationChat{
+					TextGenerationChat: &modelPB.TextGenerationChatOutput{
+						Text: "",
+					},
+				},
+			})
+		}
+		return textGenerationChatOutputs, nil
+	default:
+		postprocessResponse, err = postProcessUnspecifiedTask(inferResponse, modelMetadata.Outputs)
+		if err != nil {
+			return nil, fmt.Errorf("unable to post-process unspecified output: %w", err)
+		}
+		outputs := postprocessResponse.([]BatchUnspecifiedTaskOutputs)
+		var rawOutputs []*modelPB.TaskOutput
+		if len(outputs) == 0 {
+			return []*modelPB.TaskOutput{}, nil
+		}
+		deserializedOutputs := outputs[0].SerializedOutputs
+		for i := range deserializedOutputs {
+			var singleImageOutput []*structpb.Struct
+
+			for _, output := range outputs {
+				unspecifiedOutput := SingleOutputUnspecifiedTaskOutput{
+					Name:     output.Name,
+					Shape:    output.Shape,
+					DataType: output.DataType,
+					Data:     output.SerializedOutputs[i],
+				}
+
+				var mapOutput map[string]any
+				b, err := json.Marshal(unspecifiedOutput)
+				if err != nil {
+					return nil, err
+				}
+				if err := json.Unmarshal(b, &mapOutput); err != nil {
+					return nil, err
+				}
+				utils.ConvertAllJSONKeySnakeCase(mapOutput)
+
+				b, err = json.Marshal(mapOutput)
+				if err != nil {
+					return nil, err
+				}
+				var structData = &structpb.Struct{}
+				err = protojson.Unmarshal(b, structData)
+
+				if err != nil {
+					return nil, err
+				}
+				singleImageOutput = append(singleImageOutput, structData)
+			}
+
+			rawOutputs = append(rawOutputs, &modelPB.TaskOutput{
+				Output: &modelPB.TaskOutput_Unspecified{
+					Unspecified: &modelPB.UnspecifiedOutput{
+						RawOutputs: singleImageOutput,
+					},
+				},
+			})
+		}
+		if len(rawOutputs) == 0 {
+			rawOutputs = append(rawOutputs, &modelPB.TaskOutput{
+				Output: &modelPB.TaskOutput_Unspecified{
+					Unspecified: &modelPB.UnspecifiedOutput{
+						RawOutputs: []*structpb.Struct{},
+					},
+				},
+			})
+		}
+		return rawOutputs, nil
 	}
 }
