+++ pkg/ray/ray.go
@@ -3,12 +3,10 @@ package ray
 import (
 	"bytes"
 	"context"
-	"encoding/binary"
 	"encoding/json"
 	"fmt"
 	"io"
 	"log"
-	"math"
 	"net/http"
 	"os"
 	"path"
@@ -221,295 +219,15 @@ func (r *ray) ModelInferRequest(ctx context.Context, task commonPB.Task, inferIn
 
 	ctx = metadata.AppendToOutgoingContext(ctx, "application", applicationMetadatValue)
 
-	// Create request input tensors
-	var inferInputs []*rayserver.InferTensor
-	for i := 0; i < len(modelMetadata.Inputs); i++ {
-		switch task {
-		case commonPB.Task_TASK_IMAGE_TO_IMAGE,
-			commonPB.Task_TASK_TEXT_TO_IMAGE:
-			inferInputs = append(inferInputs, &rayserver.InferTensor{
-				Name:     modelMetadata.Inputs[i].Name,
-				Datatype: modelMetadata.Inputs[i].Datatype,
-				Shape:    []int64{1},
-			})
-		case commonPB.Task_TASK_VISUAL_QUESTION_ANSWERING,
-			commonPB.Task_TASK_TEXT_GENERATION_CHAT,
-			commonPB.Task_TASK_TEXT_GENERATION:
-			var inputShape []int64
-			inputShape = []int64{1}
+	modelInferRequest := PreProcess(modelName, version, inferInput, task, modelMetadata)
 
-			inferInputs = append(inferInputs, &rayserver.InferTensor{
-				Name:     modelMetadata.Inputs[i].Name,
-				Datatype: modelMetadata.Inputs[i].Datatype,
-				Shape:    inputShape,
-			})
-		case commonPB.Task_TASK_CLASSIFICATION,
-			commonPB.Task_TASK_DETECTION,
-			commonPB.Task_TASK_KEYPOINT,
-			commonPB.Task_TASK_OCR,
-			commonPB.Task_TASK_INSTANCE_SEGMENTATION,
-			commonPB.Task_TASK_SEMANTIC_SEGMENTATION:
-			batchSize := int64(len(inferInput.([][]byte)))
-			inferInputs = append(inferInputs, &rayserver.InferTensor{
-				Name:     modelMetadata.Inputs[i].Name,
-				Datatype: modelMetadata.Inputs[i].Datatype,
-				Shape:    []int64{batchSize, 1},
-			})
-		default:
-			batchSize := int64(len(inferInput.([][]byte)))
-			inferInputs = append(inferInputs, &rayserver.InferTensor{
-				Name:     modelMetadata.Inputs[i].Name,
-				Datatype: modelMetadata.Inputs[i].Datatype,
-				Shape:    []int64{batchSize, 1},
-			})
-		}
-	}
-
-	// Create request input output tensors
-	var inferOutputs []*rayserver.RayServiceCallRequest_InferRequestedOutputTensor
-	for i := 0; i < len(modelMetadata.Outputs); i++ {
-		switch task {
-		case commonPB.Task_TASK_CLASSIFICATION:
-			inferOutputs = append(inferOutputs, &rayserver.RayServiceCallRequest_InferRequestedOutputTensor{
-				Name: modelMetadata.Outputs[i].Name,
-			})
-		case commonPB.Task_TASK_DETECTION:
-			inferOutputs = append(inferOutputs, &rayserver.RayServiceCallRequest_InferRequestedOutputTensor{
-				Name: modelMetadata.Outputs[i].Name,
-			})
-		default:
-			inferOutputs = append(inferOutputs, &rayserver.RayServiceCallRequest_InferRequestedOutputTensor{
-				Name: modelMetadata.Outputs[i].Name,
-			})
-		}
-	}
-
-	// Create inference request for specific model/version
-	modelInferRequest := rayserver.RayServiceCallRequest{
-		ModelName:    modelName,
-		ModelVersion: version,
-		Inputs:       inferInputs,
-		Outputs:      inferOutputs,
-	}
-
-	switch task {
-	case commonPB.Task_TASK_TEXT_TO_IMAGE:
-		textToImageInput := inferInput.(*TextToImageInput)
-		samples := make([]byte, 4)
-		binary.LittleEndian.PutUint32(samples, uint32(textToImageInput.Samples))
-		steps := make([]byte, 4)
-		binary.LittleEndian.PutUint32(steps, uint32(textToImageInput.Steps))
-		guidanceScale := make([]byte, 4)
-		binary.LittleEndian.PutUint32(guidanceScale, math.Float32bits(textToImageInput.CfgScale)) // Fixed value.
-		seed := make([]byte, 8)
-		binary.LittleEndian.PutUint64(seed, uint64(textToImageInput.Seed))
-		modelInferRequest.RawInputContents = append(
-			modelInferRequest.RawInputContents,
-			SerializeBytesTensor([][]byte{[]byte(textToImageInput.Prompt)}),
-			SerializeBytesTensor([][]byte{[]byte("NONE")}),
-			SerializeBytesTensor([][]byte{[]byte(textToImageInput.PromptImage)}),
-			samples,
-			SerializeBytesTensor([][]byte{[]byte("DPMSolverMultistepScheduler")}), // Fixed value
-			steps,
-			guidanceScale,
-			seed,
-			SerializeBytesTensor([][]byte{[]byte(textToImageInput.ExtraParams)}),
-		)
-	case commonPB.Task_TASK_IMAGE_TO_IMAGE:
-		imageToImageInput := inferInput.(*ImageToImageInput)
-		samples := make([]byte, 4)
-		binary.LittleEndian.PutUint32(samples, uint32(imageToImageInput.Samples))
-		steps := make([]byte, 4)
-		binary.LittleEndian.PutUint32(steps, uint32(imageToImageInput.Steps))
-		guidanceScale := make([]byte, 4)
-		binary.LittleEndian.PutUint32(guidanceScale, math.Float32bits(imageToImageInput.CfgScale)) // Fixed value.
-		seed := make([]byte, 8)
-		binary.LittleEndian.PutUint64(seed, uint64(imageToImageInput.Seed))
-		modelInferRequest.RawInputContents = append(
-			modelInferRequest.RawInputContents,
-			SerializeBytesTensor([][]byte{[]byte(imageToImageInput.Prompt)}),
-			SerializeBytesTensor([][]byte{[]byte("NONE")}),
-			SerializeBytesTensor([][]byte{[]byte(imageToImageInput.PromptImage)}),
-			samples,
-			SerializeBytesTensor([][]byte{[]byte("DPMSolverMultistepScheduler")}), // Fixed value,
-			steps,
-			guidanceScale,
-			seed,
-			SerializeBytesTensor([][]byte{[]byte(imageToImageInput.ExtraParams)}),
-		)
-	case commonPB.Task_TASK_VISUAL_QUESTION_ANSWERING:
-		visualQUestionAnsweringInput := inferInput.(*VisualQuestionAnsweringInput)
-		maxNewToken := make([]byte, 4)
-		binary.LittleEndian.PutUint32(maxNewToken, uint32(visualQUestionAnsweringInput.MaxNewTokens))
-		temperature := make([]byte, 4)
-		binary.LittleEndian.PutUint32(temperature, math.Float32bits(visualQUestionAnsweringInput.Temperature))
-		topK := make([]byte, 4)
-		binary.LittleEndian.PutUint32(topK, uint32(visualQUestionAnsweringInput.TopK))
-		seed := make([]byte, 8)
-		binary.LittleEndian.PutUint64(seed, uint64(visualQUestionAnsweringInput.Seed))
-		modelInferRequest.RawInputContents = append(
-			modelInferRequest.RawInputContents,
-			SerializeBytesTensor([][]byte{[]byte(visualQUestionAnsweringInput.Prompt)}),
-			SerializeBytesTensor([][]byte{[]byte(visualQUestionAnsweringInput.PromptImages)}),
-			SerializeBytesTensor([][]byte{[]byte(visualQUestionAnsweringInput.ChatHistory)}),
-			SerializeBytesTensor([][]byte{[]byte(visualQUestionAnsweringInput.SystemMessage)}),
-			maxNewToken,
-			temperature,
-			topK,
-			seed,
-			SerializeBytesTensor([][]byte{[]byte(visualQUestionAnsweringInput.ExtraParams)}),
-		)
-	case commonPB.Task_TASK_TEXT_GENERATION_CHAT:
-		textGenerationChatInput := inferInput.(*TextGenerationChatInput)
-		maxNewToken := make([]byte, 4)
-		binary.LittleEndian.PutUint32(maxNewToken, uint32(textGenerationChatInput.MaxNewTokens))
-		temperature := make([]byte, 4)
-		binary.LittleEndian.PutUint32(temperature, math.Float32bits(textGenerationChatInput.Temperature))
-		topK := make([]byte, 4)
-		binary.LittleEndian.PutUint32(topK, uint32(textGenerationChatInput.TopK))
-		seed := make([]byte, 8)
-		binary.LittleEndian.PutUint64(seed, uint64(textGenerationChatInput.Seed))
-		modelInferRequest.RawInputContents = append(
-			modelInferRequest.RawInputContents,
-			SerializeBytesTensor([][]byte{[]byte(textGenerationChatInput.Prompt)}),
-			SerializeBytesTensor([][]byte{[]byte(textGenerationChatInput.PromptImages)}),
-			SerializeBytesTensor([][]byte{[]byte(textGenerationChatInput.ChatHistory)}),
-			SerializeBytesTensor([][]byte{[]byte(textGenerationChatInput.SystemMessage)}),
-			maxNewToken,
-			temperature,
-			topK,
-			seed,
-			SerializeBytesTensor([][]byte{[]byte(textGenerationChatInput.ExtraParams)}),
-		)
-	case commonPB.Task_TASK_TEXT_GENERATION:
-		textGenerationInput := inferInput.(*TextGenerationInput)
-		maxNewToken := make([]byte, 4)
-		binary.LittleEndian.PutUint32(maxNewToken, uint32(textGenerationInput.MaxNewTokens))
-		temperature := make([]byte, 4)
-		binary.LittleEndian.PutUint32(temperature, math.Float32bits(textGenerationInput.Temperature))
-		topK := make([]byte, 4)
-		binary.LittleEndian.PutUint32(topK, uint32(textGenerationInput.TopK))
-		seed := make([]byte, 8)
-		binary.LittleEndian.PutUint64(seed, uint64(textGenerationInput.Seed))
-		modelInferRequest.RawInputContents = append(
-			modelInferRequest.RawInputContents,
-			SerializeBytesTensor([][]byte{[]byte(textGenerationInput.Prompt)}),
-			SerializeBytesTensor([][]byte{[]byte(textGenerationInput.PromptImages)}),
-			SerializeBytesTensor([][]byte{[]byte(textGenerationInput.ChatHistory)}),
-			SerializeBytesTensor([][]byte{[]byte(textGenerationInput.SystemMessage)}),
-			maxNewToken,
-			temperature,
-			topK,
-			seed,
-			SerializeBytesTensor([][]byte{[]byte(textGenerationInput.ExtraParams)}),
-		)
-	case commonPB.Task_TASK_CLASSIFICATION,
-		commonPB.Task_TASK_DETECTION,
-		commonPB.Task_TASK_KEYPOINT,
-		commonPB.Task_TASK_OCR,
-		commonPB.Task_TASK_INSTANCE_SEGMENTATION,
-		commonPB.Task_TASK_SEMANTIC_SEGMENTATION:
-		modelInferRequest.RawInputContents = append(modelInferRequest.RawInputContents, SerializeBytesTensor(inferInput.([][]byte)))
-	default:
-		modelInferRequest.RawInputContents = append(modelInferRequest.RawInputContents, SerializeBytesTensor(inferInput.([][]byte)))
-	}
-
-	// Submit inference request to server
-	modelInferResponse, err := r.rayClient.XCall__(ctx, &modelInferRequest)
+	modelInferResponse, err := r.rayClient.XCall__(ctx, modelInferRequest)
 	if err != nil {
 		logger.Error(fmt.Sprintf("Error processing InferRequest: %s", err.Error()))
 		return &rayserver.RayServiceCallResponse{}, err
 	}
 
 	return modelInferResponse, nil
-}
-
-func PostProcess(inferResponse *rayserver.RayServiceCallResponse, modelMetadata *rayserver.ModelMetadataResponse, task commonPB.Task) (any, error) {
-	var (
-		outputs any
-		err     error
-	)
-
-	switch task {
-	case commonPB.Task_TASK_CLASSIFICATION:
-		outputs, err = postProcessClassification(inferResponse, modelMetadata.Outputs[0].Name)
-		if err != nil {
-			return nil, fmt.Errorf("unable to post-process classification output: %w", err)
-		}
-	case commonPB.Task_TASK_DETECTION:
-		if len(modelMetadata.Outputs) < 2 {
-			return nil, fmt.Errorf("wrong output format of detection task")
-		}
-		outputs, err = postProcessDetection(inferResponse, modelMetadata.Outputs[0].Name, modelMetadata.Outputs[1].Name)
-		if err != nil {
-			return nil, fmt.Errorf("unable to post-process detection output: %w", err)
-		}
-	case commonPB.Task_TASK_KEYPOINT:
-		if len(modelMetadata.Outputs) < 3 {
-			return nil, fmt.Errorf("wrong output format of keypoint detection task")
-		}
-		outputs, err = postProcessKeypoint(inferResponse, modelMetadata.Outputs[0].Name, modelMetadata.Outputs[1].Name, modelMetadata.Outputs[2].Name)
-		if err != nil {
-			return nil, fmt.Errorf("unable to post-process keypoint output: %w", err)
-		}
-	case commonPB.Task_TASK_OCR:
-		if len(modelMetadata.Outputs) < 2 {
-			return nil, fmt.Errorf("wrong output format of OCR task")
-		}
-		switch len(modelMetadata.Outputs) {
-		case 2:
-			outputs, err = postProcessOcrWithoutScore(inferResponse, modelMetadata.Outputs[0].Name, modelMetadata.Outputs[1].Name)
-			if err != nil {
-				return nil, fmt.Errorf("unable to post-process detection output: %w", err)
-			}
-		case 3:
-			outputs, err = postProcessOcrWithScore(inferResponse, modelMetadata.Outputs[0].Name, modelMetadata.Outputs[1].Name, modelMetadata.Outputs[2].Name)
-			if err != nil {
-				return nil, fmt.Errorf("unable to post-process detection output: %w", err)
-			}
-		}
-
-	case commonPB.Task_TASK_INSTANCE_SEGMENTATION:
-		if len(modelMetadata.Outputs) < 4 {
-			return nil, fmt.Errorf("wrong output format of instance segmentation task")
-		}
-		outputs, err = postProcessInstanceSegmentation(inferResponse, modelMetadata.Outputs[0].Name, modelMetadata.Outputs[1].Name, modelMetadata.Outputs[2].Name, modelMetadata.Outputs[3].Name)
-		if err != nil {
-			return nil, fmt.Errorf("unable to post-process instance segmentation output: %w", err)
-		}
-
-	case commonPB.Task_TASK_SEMANTIC_SEGMENTATION:
-		if len(modelMetadata.Outputs) < 2 {
-			return nil, fmt.Errorf("wrong output format of semantic segmentation task")
-		}
-		outputs, err = postProcessSemanticSegmentation(inferResponse, modelMetadata.Outputs[0].Name, modelMetadata.Outputs[1].Name)
-		if err != nil {
-			return nil, fmt.Errorf("unable to post-process semantic segmentation output: %w", err)
-		}
-
-	case commonPB.Task_TASK_IMAGE_TO_IMAGE,
-		commonPB.Task_TASK_TEXT_TO_IMAGE:
-		outputs, err = postProcessTextToImage(inferResponse, modelMetadata.Outputs[0].Name, task)
-		if err != nil {
-			return nil, fmt.Errorf("unable to post-process text to image output: %w", err)
-		}
-
-	case commonPB.Task_TASK_VISUAL_QUESTION_ANSWERING,
-		commonPB.Task_TASK_TEXT_GENERATION_CHAT,
-		commonPB.Task_TASK_TEXT_GENERATION:
-		outputs, err = postProcessTextGeneration(inferResponse, modelMetadata.Outputs[0].Name, task)
-		if err != nil {
-			return nil, fmt.Errorf("unable to post-process text to text output: %w", err)
-		}
-
-	default:
-		outputs, err = postProcessUnspecifiedTask(inferResponse, modelMetadata.Outputs)
-		if err != nil {
-			return nil, fmt.Errorf("unable to post-process unspecified output: %w", err)
-		}
-	}
-
-	return outputs, nil
 }
 
 func (r *ray) UpdateContainerizedModel(ctx context.Context, modelName string, userID string, imageName string, version string, hardware string, isDeploy bool) error {
