+++ pkg/service/worker.go
@@ -4,10 +4,15 @@ import (
 	"context"
 	"fmt"
 
+	workflowpb "go.temporal.io/api/workflow/v1"
+	rpcStatus "google.golang.org/genproto/googleapis/rpc/status"
+
 	"cloud.google.com/go/longrunning/autogen/longrunningpb"
 	"go.temporal.io/api/enums/v1"
-	"google.golang.org/genproto/googleapis/rpc/status"
+	"google.golang.org/protobuf/encoding/protojson"
 	"google.golang.org/protobuf/types/known/anypb"
+
+	modelPB "github.com/instill-ai/protogen-go/model/model/v1alpha"
 )
 
 func (s *service) GetOperation(ctx context.Context, workflowID string) (*longrunningpb.Operation, error) {
@@ -16,36 +21,38 @@ func (s *service) GetOperation(ctx context.Context, workflowID string) (*longrun
 		return nil, err
 	}
 
-	workflowExecutionInfo := workflowExecutionRes.WorkflowExecutionInfo
+	return s.getOperationFromWorkflowInfo(ctx, workflowExecutionRes.WorkflowExecutionInfo)
+}
 
+func (s *service) getOperationFromWorkflowInfo(ctx context.Context, workflowExecutionInfo *workflowpb.WorkflowExecutionInfo) (*longrunningpb.Operation, error) {
 	operation := longrunningpb.Operation{}
 
 	switch workflowExecutionInfo.Status {
 	case enums.WORKFLOW_EXECUTION_STATUS_COMPLETED:
-		var result error
-		workflowRun := s.temporalClient.GetWorkflow(ctx, workflowID, "")
-		err = workflowRun.Get(ctx, &result)
+
+		triggerModelResp := &modelPB.TriggerUserModelResponse{}
+
+		blobRedisKey := fmt.Sprintf("async_model_response:%s", workflowExecutionInfo.Execution.WorkflowId)
+		blob, err := s.redisClient.Get(ctx, blobRedisKey).Bytes()
 		if err != nil {
 			return nil, err
 		}
-		if result != nil {
-			operation = longrunningpb.Operation{
-				Done: true,
-				Result: &longrunningpb.Operation_Error{
-					Error: &status.Status{
-						Code:    int32(enums.WORKFLOW_EXECUTION_STATUS_FAILED),
-						Details: []*anypb.Any{},
-						Message: result.Error(),
-					},
-				},
-			}
-		} else {
-			operation = longrunningpb.Operation{
-				Done: true,
-				Result: &longrunningpb.Operation_Response{
-					Response: &anypb.Any{},
-				},
-			}
+
+		err = protojson.Unmarshal(blob, triggerModelResp)
+		if err != nil {
+			return nil, err
+		}
+
+		resp, err := anypb.New(triggerModelResp)
+		if err != nil {
+			return nil, err
+		}
+		resp.TypeUrl = "buf.build/instill-ai/protobufs/model.model.v1alpha.TriggerUserModelResponse"
+		operation = longrunningpb.Operation{
+			Done: true,
+			Result: &longrunningpb.Operation_Response{
+				Response: resp,
+			},
 		}
 	case enums.WORKFLOW_EXECUTION_STATUS_RUNNING:
 	case enums.WORKFLOW_EXECUTION_STATUS_CONTINUED_AS_NEW:
@@ -59,7 +66,7 @@ func (s *service) GetOperation(ctx context.Context, workflowID string) (*longrun
 		operation = longrunningpb.Operation{
 			Done: true,
 			Result: &longrunningpb.Operation_Error{
-				Error: &status.Status{
+				Error: &rpcStatus.Status{
 					Code:    int32(workflowExecutionInfo.Status),
 					Details: []*anypb.Any{},
 					Message: "",
