+++ pkg/worker/worker.go
@@ -11,19 +11,13 @@ import (
 	"github.com/instill-ai/model-backend/pkg/repository"
 )
 
-// Namespace is the Temporal namespace for model-backend
-const Namespace = "model-backend"
-
 // TaskQueue is the Temporal task queue name for model-backend
 const TaskQueue = "model-backend"
 
 // Worker interface
 type Worker interface {
-	DeployModelWorkflow(ctx workflow.Context, param *ModelParams) error
-	DeployModelActivity(ctx context.Context, param *ModelParams) error
-	UnDeployModelWorkflow(ctx workflow.Context, param *ModelParams) error
-	UnDeployModelActivity(ctx context.Context, param *ModelParams) error
-	CreateModelWorkflow(ctx workflow.Context, param *ModelParams) error
+	TriggerModelWorkflow(ctx workflow.Context, param *TriggerModelWorkflowRequest) (*TriggerModelWorkflowResponse, error)
+	TriggerModelActivity(ctx context.Context, param *TriggerModelActivityRequest) (*TriggerModelActivityResponse, error)
 }
 
 // worker represents resources required to run Temporal workflow and activity
